kibi

Function stdin

pub fn stdin() -> Result<impl BufRead>

Construct and lock a new handle to the standard input of the current process.

Errors

This function always returns Ok(…). The return type is a Result for compatibility with other platforms.