kibi

Struct Config

pub struct Config {
    pub tab_stop: NonZeroUsize,
    pub quit_times: usize,
    pub message_dur: Duration,
    pub show_line_num: bool,
}

The global Kibi configuration.

Fields
tab_stop: NonZeroUsize

The size of a tab. Must be > 0.

quit_times: usize

The number of confirmations needed before quitting, when changes have been made since the file was last changed.

message_dur: Duration

The duration for which messages are shown in the status bar.

show_line_num: bool

Whether to display line numbers.

Implementations

impl Config

pub fn load() -> Self

Load the configuration, potentially overridden using config.ini files that can be located in the following directories:

On Linux, macOS, and other *nix systems:

/etc/kibi (system-wide configuration).

$XDG_CONFIG_HOME/kibi if environment variable $XDG_CONFIG_HOME is defined, $HOME/.config/kibi otherwise (user-level configuration).

On Windows:

%APPDATA%\Kibi

Will print warnings to stderr if a file or line cannot be parsed properly.